+++ Example/Segment-Facebook/SEGAppDelegate.m
@@ -7,7 +7,11 @@
 //
 
 #import "SEGAppDelegate.h"
+#if defined(__has_include) && __has_include(<Analytics/SEGAnalytics.h>)
 #import <Analytics/SEGAnalytics.h>
+#else
+#import <Segment/SEGAnalytics.h>
+#endif
 #import <Segment-Facebook-App-Events/SEGFacebookAppEventsIntegrationFactory.h>
 
 @implementation SEGAppDelegate
@@ -18,8 +22,15 @@
     [SEGAnalytics debug:YES];
     SEGAnalyticsConfiguration *config = [SEGAnalyticsConfiguration configurationWithWriteKey:@"gnjyuUpq7mZYtLM76mwltoiZcDsFpnfY"];
     
-    // Add any of your bundled integrations.
-    [config use:[SEGFacebookAppEventsIntegrationFactory instance]];
+    // Init the FB integration SDK
+    SEGFacebookAppEventsIntegrationFactory *fb = [SEGFacebookAppEventsIntegrationFactory instance];
+
+    // Optional - Set the data processing options to override the default options of
+    // [['LDU'], 0, 0]
+    //    [fb setDataProcessingOptions:@[ @"UDL" ] forCountry:99 forState: 99];
+
+    // Add the bundle FB integration SDK
+    [config use:fb];
     
     [SEGAnalytics setupWithConfiguration:config];