+++ Pod/Classes/SEGFacebookAppEventsIntegrationFactory.m
@@ -19,9 +19,20 @@
     return self;
 }
 
+- (id)setDataProcessingOptions:(NSArray<NSString *> *)options forCountry:(int)country forState:(int)state
+{
+    self.dataProcessingOptions = options;
+    self.dataProcessingCountry = country;
+    self.dataProcessingState = state;
+    return self;
+}
+
 - (id<SEGIntegration>)createWithSettings:(NSDictionary *)settings forAnalytics:(SEGAnalytics *)analytics
 {
-    return [[SEGFacebookAppEventsIntegration alloc] initWithSettings:settings];
+    return [[SEGFacebookAppEventsIntegration alloc] initWithSettings:settings
+                                                    dataProcessingOptions:self.dataProcessingOptions
+                                                    dataProcessingCountry:self.dataProcessingCountry
+                                                    dataProcessingState:self.dataProcessingState];
 }
 
 - (NSString *)key